move build_task_config into task.py
